`v12` is for Adriana (#1112)
![CANDLELIGHT VIGIL IN MEMORY OF
ADRIANA](https://github.com/RobinTail/express-zod-api/assets/13189514/6f0242d5-4294-42b1-a7ad-5587cd3d4474)

Adriana dreamt of leaving Armenia to work as a model in a country where
pursuing a public-facing career would not expose her to danger. In this
dream, she would take her dog Froggie and flee to a place where she
could start a family of her own. She did not believe she was strong
enough to endure the discrimination and bigotry she would face in
Armenia, where there are no laws protecting the rights of the
transgender community.

She shared these aspirations with her friend Monica. Adriana never had a
large circle of friends. She sought “quality, not quantity in
friendship,” in Monica’s words. She also shared her fears with her
friend. Adriana always felt that she was in danger, but the police
dismissed her reports as products of her “imagination, not real proof.”


https://armenianweekly.com/2023/08/23/i-want-to-live-trans-woman-murdered-in-armenia/

28-year-old trans woman Adriana was brutally murdered in her own
apartment in Yerevan. The incident took place on August 20, 2023, after
the murder, the apartment was burned. Adriana said a week ago that she
was being followed and threatened.
A criminal proceeding was initiated in connection with the case,
according to the features of Article 155, Part 1 and Article 264, Part
2, Clauses 2 and 6 of the RA Criminal Code.
26-year-old suspect from Armavir region was found at the Bavra border
checkpoint, arrested, and transferred to the Central Department of the
Ministry of Internal Affairs. He was presented to the body conducting
the proceedings and questioned about the circumstances of the incident,
during which he admitted that he had committed the murder, after which
he had set the apartment on fire and tried to cross the state border.


http://rightsidengo.com/news/public-statement-trans-woman-brutally-killed-in-yerevan/

The cases of hatred and violence against Trans and LGBIQ people do not
find an appropriate response, which has already led to the formation of
an atmosphere of impunity. Unfortunately, even at this tragic moment,
when the LGBTIQ community and the relatives of the brutally murdered
trans woman were sharing their pain and grief by holding a symbolic
candlelight vigil, a group of people dared to attack and throw eggs,
bottles, stones, and sand at the gathered people, causing anxiety and
panic among persons who already were in a tough psychological situation.


http://rightsidengo.com/news/during-the-candlelight-vigil-in-memory-of-adriana-a-group-of-people-threw-eggs-bottles-and-stones-at-the-crowd/

---------------------

- #1111 
- Closes #1059 
-
https://github.com/RobinTail/express-zod-api/blob/970efbb8510797e8d3e3ecdc2f33ebb8191d59c4/tsup.config.ts#L7
- #1113 
- Closes #1114 
- #1115 
-
https://github.com/RobinTail/express-zod-api/blob/970efbb8510797e8d3e3ecdc2f33ebb8191d59c4/src/integration.ts#L308
- #1117 
- #1118 
- #1119 
-
https://github.com/RobinTail/express-zod-api/blob/970efbb8510797e8d3e3ecdc2f33ebb8191d59c4/tests/unit/documentation.spec.ts#L497
-
https://github.com/RobinTail/express-zod-api/blob/970efbb8510797e8d3e3ecdc2f33ebb8191d59c4/src/metadata.ts#L23-L28
- #1120 
- #1136