`v18` is for Victoria (#1639)
# Dedication


![image](https://github.com/RobinTail/express-zod-api/assets/13189514/7ce8a788-6254-4d01-8e7a-6dc7a82a2b05)

**Victoria Carmen White**, a black trans woman, was murdered in New
Jersey, US, on 12 September 2010. Her boyfriend was brought to trial for
her death, but he was acquitted and blamed another man for the murder.
The gun used in the killing was never found.

Transgender women suffer too frequently from transphobic violence and
cruelty, being the less protected social group. I'd like to raise an
awareness of this problem. Humans should be creators — not killers. But
most importantly, I want every transgender girl to have an opportunity
to create applications quickly and, in general, learn to write code
easily in order to receive job offers and leave dangerously transphobic
territories for more favorable and civilized ones, and live happily
there. Protect transgender women.

---

# Version 18

## Description

This version is focused on making the quick start with Express Zod API
even faster by replacing Winston with a built-in console logger having
the same basic functionality needed to begin the development. Migration
should be easy in most cases.

## What's included

- #1636 
- #1640, later overridden by:
- #1654 
- #1671