Feat: Support async functions by `.addOptions()` (#1753)
I read the documentation and realized that it does not make sense for
`.addOptions` just to proxy the static data.
For that purpose one could just `import` the previously assigned
`const`. That's it.
What does make sense is to accept an async function by that method.

Static option will be deprecated and removed in v19.