v23 is for Sonia (#2357)
![image](https://github.com/user-attachments/assets/124f41b0-a43f-40f1-8256-2ff7e7aed6f8)

Sonia Rescalvo Zafra was a Spanish trans woman who was murdered by a
group of neo-Nazis in the _Parc de la Ciutadella_ in Barcelona on 6
October 1991. Her murder is notable for being the first known case in
Spain of a person being murdered for being transgender.

Sonia was born in _Cuenca_, Spain in 1946. She moved to Barcelona at the
age of 16 and began working as a showgirl in theatres on the _Avinguda
del Paral·lel_. Later in life, she fell on hard times and became
homeless and addicted to drugs, turning to sex work to survive.

Sonia and a her friend, Doris Romero, were sleeping rough at a bandstand
in Barcelona's Parc _de la Ciutadella_ in the early hours of 6 October
1991 when they were abruptly attacked by a group of six teenage
skinheads from the _Boixos Nois_, who beat them with a broomstick and
repeatedly kicked them in the head while wearing steel-toe boots. Sonia
was killed in the attack, while Doris was seriously injured and left
unconscious. The perpetrators went on to attack three other people
sleeping near the bandstand. One of the three other victims, Miguel
Pérez Barreiros, was able to contact police after the attack, despite
being left completely blind from his injuries. When police discovered
Sonia's body, she was so badly bruised that media reports initially
misidentified her as a black woman.

In June 1994, six of the suspects were convicted of participating in the
attack and given prison sentences ranging from 23 to 50 years, while the
seventh, Lozano, was fined 100,000 pesetas for knowing of the murder and
not reporting it. In 1996, the attackers' sentences were reduced by the
Supreme Court of Spain. All six have since been released from prison.

https://en.wikipedia.org/wiki/Murder_of_Sonia_Rescalvo_Zafra

Transgender women suffer too frequently from transphobic violence and
cruelty, being the less protected social group. I'd like to raise an
awareness of this problem. Humans should be creators — not killers. But
most importantly, I want every transgender girl to have an opportunity
to create applications quickly and, in general, learn to write code
easily in order to receive job offers and leave dangerously transphobic
territories for more favorable and civilized ones, and live happily
there. Protect transgender women.

---------------------------------

This version of the framework aims to improve its maintainability and
perform several corrections. Since Express v5.1 finally tagged `latest`,
the support of Express 4 is dropped in order to take advantage of the
better support of asynchronous handlers. The default value for
`wrongMethodBehavior` changed from `404` to `405` as the more suitable
one for response in that case. Removed `errorHandler` option from
`testMiddleware()` to avoid confusion with config option having same
name. Several public methods and properties intended for internal usage
are removed from types declaration so that potential changes to them
would not be considered breaking, enabling more improvements in the
future.

- #2358 
  - #2359 
- #2372  
- #2387 
- #2398 
- #2401 
  - #2402 
- #2475 
  - #2524 
  - #2539 
- #2522